Extract shared bootstrap for server binary and desktop app
The tauri plugin duplicated the entire ~150-line composition root from
bin/tanoshi.rs, and the copies had drifted: the desktop app swallowed
database and extension-load errors silently, never wired the configured
notifiers (Telegram/Pushover/Gotify were dead on desktop), hardcoded
create_database instead of honoring config, and silently disabled
trackers on missing base_url where the server fails fast with a clear
message.

Move the wiring into tanoshi::application::bootstrap (gated on the
server feature) returning an App with the ready ServerBuilder, worker
handles, pool, and optional Telegram bot future. Both frontends now
share one code path; the binary keeps its listen address and select
loop, the desktop app keeps its random localhost port and now logs
startup failures instead of dying silently. Also replace the unwraps
in the tauri window setup with an expect and a logged error.
