+++ crates/tanoshi/src/application/mod.rs
@@ -1,1 +1,3 @@
+#[cfg(feature = "server")]
+pub mod bootstrap;
 pub mod worker;
